+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Ensured compatibility with PHP 8
   * Added support for union types
   * Added support for mixed type hint
+  * Added support for static return type hint
 * Ensured compatibility with PHPUnit 9.2 and later
 
 ## 6.1.0 (2020-03-02)